xml/parsing/manakai/attlists-1.dat

#data
<!DOCTYPE a[
<!ATTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ATTLIST c>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;1;empty attlist decl;;c;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|   <!ELEMENT c >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST bcd>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;bcd;w
#document
| <!DOCTYPE a>
|   <!ELEMENT bcd >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST \u0000bcd>
]><a/>
#errors escaped
1;1;no XML decl;;;s
2;1;empty attlist decl;;\uFFFDbcd;w
2;11;NULL;;;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT \uFFFDbcd >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST a\u0000bcd>
]><a/>
#errors escaped
1;1;no XML decl;;;s
2;1;empty attlist decl;;a\uFFFDbcd;w
2;12;NULL;;;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT a\uFFFDbcd >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b >
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b\u0020
#errors
1;1;no XML decl;;;s
2;13;parser:EOF;;;m
2;13;parser:EOF;;;m
2;13;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST bcd >
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;bcd;w
#document
| <!DOCTYPE a>
|   <!ELEMENT bcd >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLISTb>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
2;10;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
#errors
1;1;no XML decl;;;s
2;12;parser:EOF;;;m
2;12;parser:EOF;;;m
2;12;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST >
<!ATTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
2;11;no md name;;;m
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST>
<!ATTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
3;1;empty attlist decl;;b;w
2;10;no md name;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b>
<!ATTLIST\u0020
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;11;parser:EOF;;;m
3;11;parser:EOF;;;m
3;11;no root element;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ATTLIST
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;10;parser:EOF;;;m
3;10;parser:EOF;;;m
3;10;no root element;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >

#data
<!DOCTYPE a[
<!ATTLIS>
<!ATTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
2;9;bogus comment;;;m
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATT c>
<!ATTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
2;6;bogus comment;;;m
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!attlist c>
<!ATTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
2;3;keyword-wrong-case;;attlist;m
2;1;empty attlist decl;;c;w
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|   <!ELEMENT c >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ATTLIS
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;9;bogus comment;;;m
3;9;parser:EOF;;;m
3;9;no root element;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ATTL
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;7;bogus comment;;;m
3;7;parser:EOF;;;m
3;7;no root element;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >

#data
<!DOCTYPE a[
<!ATTLIST% c>
<!ATTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
2;11;no refc;;;m
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ATTLIST % c>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;12;no refc;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
2;14;no attr type;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c
#errors
1;1;no XML decl;;;s
2;14;parser:EOF;;;m
2;14;parser:EOF;;;m
2;14;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b cd>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
2;15;no attr type;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b cd
#errors
1;1;no XML decl;;;s
2;15;parser:EOF;;;m
2;15;parser:EOF;;;m
2;15;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c >
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
2;15;no attr type;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b c\u0020
#errors
1;1;no XML decl;;;s
2;15;parser:EOF;;;m
2;15;parser:EOF;;;m
2;15;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b cd >
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
2;16;no attr type;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b cd\u0020
#errors
1;1;no XML decl;;;s
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c t>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
2;16;no attr default;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c t
#errors
1;1;no XML decl;;;s
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c tu>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
2;17;no attr default;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c tu
#errors
1;1;no XML decl;;;s
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c t >
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
2;17;no attr default;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b c t\u0020
#errors
1;1;no XML decl;;;s
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c CDATA>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
2;20;no attr default;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c CDATA
#errors
1;1;no XML decl;;;s
2;20;parser:EOF;;;m
2;20;parser:EOF;;;m
2;20;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c t d>
]><a/>
#errors
1;1;no XML decl;;;s
2;17;unquoted attr value;;;m
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c t d
#errors
1;1;no XML decl;;;s
2;17;unquoted attr value;;;m
2;18;parser:EOF;;;m
2;18;parser:EOF;;;m
2;18;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID d>
]><a/>
#errors
1;1;no XML decl;;;s
2;18;unquoted attr value;;;m
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID d
#errors
1;1;no XML decl;;;s
2;18;unquoted attr value;;;m
2;19;parser:EOF;;;m
2;19;parser:EOF;;;m
2;19;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID "d>
]><a/>
#errors
1;1;no XML decl;;;s
3;3;lt in attr value;;;m
3;7;parser:EOF;;;m
3;7;parser:EOF;;;m
3;7;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID "d
#errors
1;1;no XML decl;;;s
2;20;parser:EOF;;;m
2;20;parser:EOF;;;m
2;20;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID "d">
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ID () EXPLICIT "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID "d"
#errors
1;1;no XML decl;;;s
2;21;parser:EOF;;;m
2;21;parser:EOF;;;m
2;21;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID "d&amp;">
]><a/>
#errors
1;1;no XML decl;;;s
2;20;entity not declared;;&amp;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ID () EXPLICIT "d&"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID 'd>
]><a/>
#errors
1;1;no XML decl;;;s
3;3;lt in attr value;;;m
3;7;parser:EOF;;;m
3;7;parser:EOF;;;m
3;7;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID 'd
#errors
1;1;no XML decl;;;s
2;20;parser:EOF;;;m
2;20;parser:EOF;;;m
2;20;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID 'd'>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ID () EXPLICIT "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID 'd'
#errors
1;1;no XML decl;;;s
2;21;parser:EOF;;;m
2;21;parser:EOF;;;m
2;21;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID 'd&amp;'>
]><a/>
#errors
1;1;no XML decl;;;s
2;20;entity not declared;;&amp;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ID () EXPLICIT "d&"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID d&amp;>
]><a/>
#errors
1;1;no XML decl;;;s
2;18;unquoted attr value;;;m
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA "d"
e CDATA "f">
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "d"
|     e CDATA () EXPLICIT "f"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA "d"e CDATA "f">
]><a/>
#errors
1;1;no XML decl;;;s
3;12;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "d"
|     e CDATA () EXPLICIT "f"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA 'd'
e CDATA "f"
>]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "d"
|     e CDATA () EXPLICIT "f"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA 'd'e CDATA "f"
>]><a/>
#errors
1;1;no XML decl;;;s
3;12;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "d"
|     e CDATA () EXPLICIT "f"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA d
e CDATA "f"
>]><a/>
#errors
1;1;no XML decl;;;s
3;9;unquoted attr value;;;m
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA"d"
>]><a/>
#errors
1;1;no XML decl;;;s
3;8;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA'd'
>]><a/>
#errors
1;1;no XML decl;;;s
3;8;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "d"
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
\u0000c CDATA "d">
]><a/>
#errors
1;1;no XML decl;;;s
3;1;NULL;;;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
|     \uFFFDc CDATA () EXPLICIT "d"
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
x\u0000c CDATA "d">
]><a/>
#errors
1;1;no XML decl;;;s
3;2;NULL;;;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
|     x\uFFFDc CDATA () EXPLICIT "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #
#errors
1;1;no XML decl;;;s
3;10;parser:EOF;;;m
3;10;parser:EOF;;;m
3;10;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
c CDATA #\u0020
#errors
1;1;no XML decl;;;s
3;10;no attr default;;;m
3;11;parser:EOF;;;m
3;11;parser:EOF;;;m
3;11;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA ##
#errors
1;1;no XML decl;;;s
3;11;parser:EOF;;;m
3;11;parser:EOF;;;m
3;11;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA ##
>]><a/>
#errors
1;1;no XML decl;;;s
3;1;unknown default type;;#;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () 0 ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA#REQUIRED
>]><a/>
#errors
1;1;no XML decl;;;s
3;8;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () REQUIRED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA#IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
3;8;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA#FIXED "d"
>]><a/>
#errors
1;1;no XML decl;;;s
3;8;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () FIXED "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #FIXED"d"
>]><a/>
#errors
1;1;no XML decl;;;s
3;15;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () FIXED "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #IMPLIED
d CDATA #IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () IMPLIED ""
|     d CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #REQUIRED
d CDATA #IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () REQUIRED ""
|     d CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #FIXED "x"
d CDATA #IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () FIXED "x"
|     d CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #UNKNOWN
d CDATA #IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
3;1;unknown default type;;UNKNOWN;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () 0 ""
|     d CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c IDREF #IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c IDREF () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c IDREFS #IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c IDREFS () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NMTOKEN #IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NMTOKEN () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NMTOKENS #IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NMTOKENS () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c ENTITY #IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENTITY () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c ENTITIES #IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENTITIES () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NOTATION #IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NOTATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c ENUMERATION #IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
3;1;unknown declared type;;ENUMERATION;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c 0 () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c cdata #IMPLIED
>]><a/>
#errors
1;1;no XML decl;;;s
3;1;unknown declared type;;cdata;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c 0 () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #FIXED 'd'
e CDATA #IMPLIED
>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () FIXED "d"
|     e CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #FIXED'd'
e CDATA #IMPLIED
>
]><a/>
#errors
1;1;no XML decl;;;s
3;15;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () FIXED "d"
|     e CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #FIXED 'd'e CDATA #IMPLIED
>
]><a/>
#errors
1;1;no XML decl;;;s
3;19;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () FIXED "d"
|     e CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #x 'd'
e CDATA #IMPLIED
>
]><a/>
#errors
1;1;no XML decl;;;s
3;1;unknown default type;;x;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () 0 ""
|     e CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA # 'd'
e CDATA #IMPLIED
>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;10;no attr default;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c () #IMPLIED
>
]><a/>
#errors
1;1;no XML decl;;;s
3;4;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NOTATION () #IMPLIED
>
]><a/>
#errors
1;1;no XML decl;;;s
3;13;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NOTATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c ()#IMPLIED
>
]><a/>
#errors
1;1;no XML decl;;;s
3;4;xml:dtd:no group item;;;m
3;5;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NOTATION ()#IMPLIED
>
]><a/>
#errors
1;1;no XML decl;;;s
3;13;xml:dtd:no group item;;;m
3;14;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NOTATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c() #IMPLIED
>
]><a/>
#errors
1;1;no XML decl;;;s
3;2;no space between params;;;m
3;3;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NOTATION() #IMPLIED
>
]><a/>
#errors
1;1;no XML decl;;;s
3;11;no space between params;;;m
3;12;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NOTATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c () "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;4;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NOTATION () "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;13;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NOTATION () EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c ()"x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;4;xml:dtd:no group item;;;m
3;5;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NOTATION ()"x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;13;xml:dtd:no group item;;;m
3;14;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NOTATION () EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (y) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (y) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (yz) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (yz) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (y|z) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (y|z) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (y | z | abc) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (y|z|abc) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (   ABC  |DEF|G  ) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (ABC|DEF|G) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (|) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;4;xml:dtd:no group item;;;m
3;5;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (|y) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;4;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (y) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (y|) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;6;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (y) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (y||z) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;6;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (y|z) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (||) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;4;xml:dtd:no group item;;;m
3;5;xml:dtd:no group item;;;m
3;6;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (
#errors
1;1;no XML decl;;;s
3;4;parser:EOF;;;m
3;4;parser:EOF;;;m
3;4;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
c (\u0020
#errors
1;1;no XML decl;;;s
3;5;parser:EOF;;;m
3;5;parser:EOF;;;m
3;5;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x
#errors
1;1;no XML decl;;;s
3;5;parser:EOF;;;m
3;5;parser:EOF;;;m
3;5;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (xy
#errors
1;1;no XML decl;;;s
3;6;parser:EOF;;;m
3;6;parser:EOF;;;m
3;6;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
c (x\u0020
#errors
1;1;no XML decl;;;s
3;6;parser:EOF;;;m
3;6;parser:EOF;;;m
3;6;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x|
#errors
1;1;no XML decl;;;s
3;6;parser:EOF;;;m
3;6;parser:EOF;;;m
3;6;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;4;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c ( >
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;5;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (a>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;5;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (ab>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;6;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x >
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;6;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x)
#errors
1;1;no XML decl;;;s
3;6;parser:EOF;;;m
3;6;parser:EOF;;;m
3;6;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
c (x)\u0009
#errors
1;1;no XML decl;;;s
3;7;parser:EOF;;;m
3;7;parser:EOF;;;m
3;7;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x|>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;6;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x y) #IMPLIED>
]><a/>
#errors
1;1;no XML decl;;;s
3;6;xml:dtd:no group delimiter;;;m
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x y  |z) #IMPLIED>
]><a/>
#errors
1;1;no XML decl;;;s
3;6;xml:dtd:no group delimiter;;;m
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x y   z) #IMPLIED>
]><a/>
#errors
1;1;no XML decl;;;s
3;6;xml:dtd:no group delimiter;;;m
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x     y) #IMPLIED>
]><a/>
#errors
1;1;no XML decl;;;s
3;10;xml:dtd:no group delimiter;;;m
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
c (\u0000yz) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;4;NULL;;;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (\uFFFDyz) EXPLICIT "x"
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
c (x\u0000yz) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;5;NULL;;;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (x\uFFFDyz) EXPLICIT "x"
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
c (  \u0000yz) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;6;NULL;;;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (\uFFFDyz) EXPLICIT "x"
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
c (a \u0000yz) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;6;xml:dtd:no group delimiter;;;m
2;1;empty attlist decl;;b;w
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
c (a|\u0000yz) "x"
>
]><a/>
#errors
1;1;no XML decl;;;s
3;6;NULL;;;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (a|\uFFFDyz) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!attlist b>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
2;3;keyword-wrong-case;;attlist;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLISt b>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
2;3;keyword-wrong-case;;ATTLISt;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!aTTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
2;3;keyword-wrong-case;;aTTLIST;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ATTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;1;duplicate attlist decl;;b;w
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ANY>
<!ATTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b ANY>
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ELEMENT b ANY>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b ANY>
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #
#errors
1;1;no XML decl;;;s
3;10;parser:EOF;;;m
3;10;parser:EOF;;;m
3;10;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #IMPLIED
#errors
1;1;no XML decl;;;s
3;17;parser:EOF;;;m
3;17;parser:EOF;;;m
3;17;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
c CDATA #IMPLIED\u0009\u0020
#errors
1;1;no XML decl;;;s
3;19;parser:EOF;;;m
3;19;parser:EOF;;;m
3;19;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #IMPLIED
c CDATA #REQUIRED>
]><a/>
#errors
1;1;no XML decl;;;s
4;1;duplicate attrdef;;c;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #IMPLIED>
<!ATTLIST b
c CDATA #REQUIRED>
]><a/>
#errors
1;1;no XML decl;;;s
4;1;duplicate attlist decl;;b;w
5;1;duplicate attrdef;;c;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c CDATA #IMPLIED>
%x;
<!ATTLIST d e CDATA #IMPLIED>
]><a/>
#errors
1;1;no XML decl;;;s
3;1;entity not declared;;%x;;m
3;1;stop processing;;;i
4;1;xml:dtd:attlist ignored;;;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () IMPLIED ""
| <{} a>

#data
<?xml version="1.0" standalone="yes"?>
<!DOCTYPE a[
<!ATTLIST b c CDATA #IMPLIED>
%x;
<!ATTLIST d e CDATA #IMPLIED>
]><a/>
#errors
4;1;entity not declared;;%x;;m
#xml-standalone
true
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () IMPLIED ""
|   <!ELEMENT d >
|     e CDATA () IMPLIED ""
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b c CDATA "ab\u0020\u0009\u000C\u000D\u000Acd">
]><a/>
#errors
1;1;no XML decl;;;s
2;26;control char;U+000C;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "ab    cd"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c CDATA "ab&#xd;&#xa;&#x9;&#x20;&#xc;">
]>
<a/>
#errors
1;1;no XML decl;;;s
2;45;invalid character reference;U+000C;;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "ab\u000d\u000a\u0009\u0020\u000c"
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b c CDATA "ab\u0000cd">
]><a/>
#errors
1;1;no XML decl;;;s
2;24;NULL;;;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "ab\uFFFDcd"
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b c CDATA 'ab\u0000cd'>
]><a/>
#errors
1;1;no XML decl;;;s
2;24;NULL;;;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "ab\uFFFDcd"
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b c CDATA ab\u0000cd>
]><a/>
#errors
1;1;no XML decl;;;s
2;21;unquoted attr value;;;m
2;1;empty attlist decl;;b;w
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b c CDATA \u0000cd>
]><a/>
#errors
1;1;no XML decl;;;s
2;21;unquoted attr value;;;m
2;1;empty attlist decl;;b;w
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>
